+++ src/models/Game.ts
@@ -1,41 +1,15 @@
-interface ResourcesList {
-	wood: number
-	maxWood: number
-	food: number
-	maxFood: number
-}
+import { BuildingsManager } from './buildings/BuildingsManager'
+import { ResourcesManager } from './resources/ResourcesManager'
+import { CatsManager } from './cats/CatsManager'
 
 export class Game {
-	private food: number
-	private wood: number
-	private maxWood: number
-	private maxFood: number
+	resourcesManager: ResourcesManager
+	buildingsManager: BuildingsManager
+	catsManager: CatsManager
 
 	constructor(initialFood: number, maxBaseWood: number, initialWood: number, maxBaseFood: number) {
-		this.food = initialFood
-		this.maxWood = maxBaseWood
-		this.wood = initialWood
-		this.maxFood = maxBaseFood
-	}
-
-	colectFood(): void {
-		if (this.food < this.maxFood) {
-			this.food += 1
-		}
-	}
-
-	colectWood(): void {
-		if (this.wood < this.maxWood) {
-			this.wood += 1
-		}
-	}
-
-	getResources(): ResourcesList {
-		return {
-			wood: this.wood,
-			maxWood: this.maxWood,
-			food: this.food,
-			maxFood: this.maxFood,
-		}
+		this.resourcesManager = new ResourcesManager(initialWood, maxBaseWood, initialFood, maxBaseFood)
+		this.buildingsManager = new BuildingsManager(this.resourcesManager)
+		this.catsManager = new CatsManager(this.resourcesManager, this.buildingsManager)
 	}
 }