+++ src/utils/i18n/locales/en/buildings.ts
@@ -1,0 +1,4 @@
+export default {
+	buyHouse: 'Buy house',
+	buyLumberCamp: 'Buy lumber camp',
+}