+++ src/utils/i18n/locales/en/common.ts
@@ -1,3 +1,9 @@
 export default {
 	title: 'Toga Idle',
+	actionsTitle: 'Actions',
+	resourcesTitle: 'Resources',
+	buildingsTitle: 'Buildings',
+	villageMapTitle: 'Map',
+	diaryTitle: 'Diary',
+  cost: 'Cost',
 }