+++ src/utils/i18n/locales/en/index.ts
@@ -1,7 +1,11 @@
 import commonMsgs from './common'
 import resourcesMsgs from './resources'
+import buildingsMsgs from './buildings'
+import catsMsgs from './cats'
 
 export default {
 	common: commonMsgs,
 	resources: resourcesMsgs,
+	buildings: buildingsMsgs,
+	cats: catsMsgs,
 }