+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-03-20
+### Added
+- Added main classes managers.
+- Created new interface.
+- Created new components (tabs, cost, tooltip, button).
+- Created new sections.
+
 ## [0.0.1] - 2024-02-27
 ### Added
 - Added base configuration for the project.